Add unsafe block to Iterators
The existing code is an exact copy-and-paste of the previous code, which
does not work in this situation. `Node<T>`'s `next` is a `Link<T>` (and
therefore `*mut Node<T>`). That type does not have `as_deref()` and
`as_deref_mut()`, instead the unsafe `as_ref()` and `as_mut()` must be
used. This change actually changes the implementation to match the
snippet from the full implementation shown in `fifth-final`.