Add copied code inside of citation
On first read, it was far from clear which part of the code was copy-pasted and which part was actually what you suggest us to try in your first-implementation of linked-list.
Personally, I failed to understand why you suddenly renamed Empty and Elem to Cons and Nil, and while you went from i32 to generic type. It become clear only when I realized that below, you moved back to an enum with the constructor names you first used and that the code came from the documentation.