Mention iterating twice in second implementation
If I were to implement tests for list in any most other languages, I consider
that I would check that iteration don't change the list, that we can iterate
twice on it. I suggest that it makes sense to add a small discussion about this
topic, explaining why it can't be done with into_iter, the tradeoff between next
giving ownership and keeping ownership of list.