Mentionning real-time queue
Reading about queue, and your sentence mentionning that it would require
traversing the sequence of nodes, I felt slightly bad. I love Okasaki's work and
feel like honesty requires to note that this is not a requirement. We can keep
the beauty of functional-programing, the simplicity of having a simple data
structure that just contains values and unmutable pointers, and still be
efficient; even in the worst-case.

As it's clearly not-relevant for the remaining of the book, I limited myself to
a single paragraph and a link.

Disclaimer: I'm linking to a part of a wikipedia article I wrote.