Add first push graphical example
In my case, I had to search the definition of mem::replace in the
rust documentation for knowing what it does and understanding why this
example works.

I think that an image would become handy to people that
are new to rust but know well the linked list data structures.